Allow customizing row attributes
Store the source row attributes, if available, to be used when rendering the table rows.

This introduces a breaking change in `rowRender`, as the signature of the `row` argument changed from `cellType[]` to `{ cells: cellType[], attributes: object }`. I have updated `Upgrading.md` with the relevant entry for the next major version.

Tested all the demos and did not find regressions. Both DOM and programmatic insertion are supported; please see demo 25 for how that looks like.

```html
<table>
    <thead>
        <tr>
            <th>Header 1</th>
            <th>Header 2</th>
        </tr>
    </thead>
    <tbody>
        <tr class="my-special-row">
            <td>Cell 1</td>
            <td>Cell 2</td>
        </tr>
    </tbody>
</table>
<script>
    new window.simpleDatatables.DataTable("table")
</script>
```